// src/types/data.interface.ts
export interface Data {
  adult: boolean;
  backdrop_path: string;
  first_air_date: any;
  release_date: any;
  genre_ids: number[];
  id: number;
  media_type: string;
  name: string;
  title: string;
  origin_country: string[];
  original_language: string;
  original_name: string;
  overview: string;
  popularity: number;
  poster_path: string;
  vote_average: number;
}